add send comment

+++ app/src/main/java/ed/maevski/testbalinasoft/domain/models/Comment.kt
@@ -1,8 +1,8 @@
 package ed.maevski.testbalinasoft.domain.models
 
 data class Comment(
-    val commentId: Long? = null,
-    val imageId: Long? = null,
+    val commentId: Int? = null,
+    val imageId: Int? = null,
     val date: Long,
     val text: String,
 )
